Use double precision in metric calculation. (#7364)

+++ src/metric/auc.h
@@ -23,59 +23,60 @@ namespace metric {
 /***********
  * ROC AUC *
  ***********/
-XGBOOST_DEVICE inline float TrapezoidArea(float x0, float x1, float y0, float y1) {
+XGBOOST_DEVICE inline double TrapezoidArea(double x0, double x1, double y0, double y1) {
   return std::abs(x0 - x1) * (y0 + y1) * 0.5f;
 }
 
 struct DeviceAUCCache;
 
-std::tuple<float, float, float>
+std::tuple<double, double, double>
 GPUBinaryROCAUC(common::Span<float const> predts, MetaInfo const &info,
                 int32_t device, std::shared_ptr<DeviceAUCCache> *p_cache);
 
-float GPUMultiClassROCAUC(common::Span<float const> predts,
-                          MetaInfo const &info, int32_t device,
-                          std::shared_ptr<DeviceAUCCache> *cache,
-                          size_t n_classes);
+double GPUMultiClassROCAUC(common::Span<float const> predts,
+                           MetaInfo const &info, int32_t device,
+                           std::shared_ptr<DeviceAUCCache> *cache,
+                           size_t n_classes);
 
-std::pair<float, uint32_t>
+std::pair<double, uint32_t>
 GPURankingAUC(common::Span<float const> predts, MetaInfo const &info,
               int32_t device, std::shared_ptr<DeviceAUCCache> *cache);
 
 /**********
  * PR AUC *
  **********/
-std::tuple<float, float, float>
+std::tuple<double, double, double>
 GPUBinaryPRAUC(common::Span<float const> predts, MetaInfo const &info,
                int32_t device, std::shared_ptr<DeviceAUCCache> *p_cache);
 
-float GPUMultiClassPRAUC(common::Span<float const> predts, MetaInfo const &info,
-                         int32_t device, std::shared_ptr<DeviceAUCCache> *cache,
-                         size_t n_classes);
+double GPUMultiClassPRAUC(common::Span<float const> predts,
+                          MetaInfo const &info, int32_t device,
+                          std::shared_ptr<DeviceAUCCache> *cache,
+                          size_t n_classes);
 
-std::pair<float, uint32_t>
+std::pair<double, uint32_t>
 GPURankingPRAUC(common::Span<float const> predts, MetaInfo const &info,
                 int32_t device, std::shared_ptr<DeviceAUCCache> *cache);
 
 namespace detail {
-XGBOOST_DEVICE inline float CalcH(float fp_a, float fp_b, float tp_a,
-                                  float tp_b) {
+XGBOOST_DEVICE inline double CalcH(double fp_a, double fp_b, double tp_a,
+                                   double tp_b) {
   return (fp_b - fp_a) / (tp_b - tp_a);
 }
 
-XGBOOST_DEVICE inline float CalcB(float fp_a, float h, float tp_a, float total_pos) {
+XGBOOST_DEVICE inline double CalcB(double fp_a, double h, double tp_a, double total_pos) {
   return (fp_a - h * tp_a) / total_pos;
 }
 
-XGBOOST_DEVICE inline float CalcA(float h) { return h + 1; }
+XGBOOST_DEVICE inline double CalcA(double h) { return h + 1; }
 
-XGBOOST_DEVICE inline float CalcDeltaPRAUC(float fp_prev, float fp,
-                                           float tp_prev, float tp,
-                                           float total_pos) {
-  float pr_prev = tp_prev / total_pos;
-  float pr = tp / total_pos;
+XGBOOST_DEVICE inline double CalcDeltaPRAUC(double fp_prev, double fp,
+                                            double tp_prev, double tp,
+                                            double total_pos) {
+  double pr_prev = tp_prev / total_pos;
+  double pr = tp / total_pos;
 
-  float h{0}, a{0}, b{0};
+  double h{0}, a{0}, b{0};
 
   if (tp == tp_prev) {
     a = 1.0;
@@ -86,7 +87,7 @@ XGBOOST_DEVICE inline float CalcDeltaPRAUC(float fp_prev, float fp,
     b = detail::CalcB(fp_prev, h, tp_prev, total_pos);
   }
 
-  float area = 0;
+  double area = 0;
   if (b != 0.0) {
     area = (pr - pr_prev -
             b / a * (std::log(a * pr + b) - std::log(a * pr_prev + b))) /
